+++ CHANGES.rst
@@ -41,6 +41,9 @@ Release History
   added to the model. See `configuration options
   <https://www.nengo.ai/nengo/nengorc.html#configuration-options>`__
   for details. (`#1532 <https://github.com/nengo/nengo/pull/1532>`__)
+- Added a ``--memory`` option for pytest that prints the total memory
+  consumed by the tests when they complete (Linux and Mac OS X only).
+  (`#640 <https://github.com/nengo/nengo/pull/640>`__)
 
 **Changed**